+++ src/common/enums/api/apiPath.ts
@@ -1,6 +1,6 @@
 enum ApiPath {
   AUTH = '/auth',
-  USERS = '/users'
+  PODCASTS = '/podcasts'
 }
 
 export { ApiPath }
