+++ src/common/enums/api/index.ts
@@ -1,0 +1,1 @@
+export { ApiPath } from './apiPath'
