+++ src/common/enums/file/index.ts
@@ -1,0 +1,1 @@
+export { ContentType } from './contentType'
