+++ src/common/enums/http/index.ts
@@ -1,0 +1,1 @@
+export { HttpMethod } from './httpMethod'
