+++ src/common/types/index.ts
@@ -1,4 +1,7 @@
 export type { RootState, AppDispatch, AsyncThunkConfig } from './app'
-export type { User, UserSignInPayload } from './user'
-export type { SignResponse } from './sign'
+export type { RecentlyPlayedEpisode } from './episode'
+export type { Genre } from './genre'
 export type { Image } from './image'
+export type { Podcast, PodcastLoadFilter, PodcastQueryPayload } from './podcast'
+export type { SignResponse } from './sign'
+export type { User, UserSignInPayload } from './user'
