+++ src/services/authApi/common/enums/authApiPath.ts
