+++ src/services/authApi/common/enums/index.ts
@@ -1,0 +1,1 @@
+export { AuthApiPath } from './authApiPath'
