--- src/services/common/enums/api/index.ts
@@ -1,3 +1,0 @@
-export { ApiPath } from './apiPath'
-export { AuthApiPath } from './authApiPath'
-export { UsersApiPath } from './usersApiPath'
