--- src/services/common/enums/api/usersApiPath.ts
@@ -1,6 +1,0 @@
-enum UsersApiPath {
-  ROOT = '/',
-  $ID = '/:id'
-}
-
-export { UsersApiPath }
