--- src/services/common/enums/file/fileExtension.ts
@@ -1,10 +1,0 @@
-enum FileExtension {
-  JPG = 'jpg',
-  PNG = 'png',
-  SVG = 'svg',
-  JPEG = 'jpeg',
-  MP3 = 'mp3',
-  WAV = 'wav'
-}
-
-export { FileExtension }
