--- src/services/common/enums/file/index.ts
@@ -1,2 +1,0 @@
-export { ContentType } from './contentType'
-export { FileExtension } from './fileExtension'
