--- src/services/common/enums/http/index.ts
@@ -1,3 +1,0 @@
-export { HttpCode } from './httpCode'
-export { HttpMethod } from './httpMethod'
-export { HttpHeader } from './httpHeader'
