--- src/services/common/enums/index.ts
@@ -1,3 +1,0 @@
-export { ApiPath, AuthApiPath, UsersApiPath } from './api'
-export { HttpCode, HttpMethod, HttpHeader } from './http'
-export { ContentType, FileExtension } from './file'
