--- src/services/common/helpers/http/getStringifiedQuery/getStringifiedQuery.ts
@@ -1,6 +1,0 @@
-import { stringify } from 'query-string'
-
-const getStringifiedQuery = (query: Record<string, string | number>): string =>
-  stringify(query)
-
-export { getStringifiedQuery }
