--- src/services/common/helpers/http/index.ts
@@ -1,1 +1,0 @@
-export { getStringifiedQuery } from './getStringifiedQuery/getStringifiedQuery'
