--- src/services/common/helpers/index.ts
@@ -1,1 +1,0 @@
-export { getStringifiedQuery } from './http'
