--- src/services/common/types/http/index.ts
@@ -1,1 +1,0 @@
-export type { HttpOptions } from './httpOptions'
