--- src/services/common/types/index.ts
@@ -1,1 +1,0 @@
-export type { HttpOptions } from './http'
