+++ src/services/http/common/helpers/getStringifiedQuery.ts
@@ -1,0 +1,7 @@
+import { stringify } from 'query-string'
+
+const getStringifiedQuery = (query: Record<string, unknown>): string => {
+  return stringify(query)
+}
+
+export { getStringifiedQuery }
