+++ src/services/http/common/types/index.ts
@@ -1,1 +1,2 @@
 export type { GetHeadersProps } from './getHeadersParams'
+export type { HttpOptions } from './httpOptions'
