+++ src/services/http/exceptions/index.ts
@@ -1,0 +1,1 @@
+export { HttpError } from './httpError'
