--- src/services/userApi/index.ts
@@ -1,1 +1,0 @@
-export { UserApi } from './userApi'
